+++ src/collector/sysmon_csv.rs
@@ -97,6 +97,10 @@ impl<T> Collector for SysmonCollector<T>
 where
     T: Serialize + TryFromSysmonRecord + Unpin + Debug + Send,
 {
+    fn kind(&self) -> RawEventKind {
+        self.protocol
+    }
+
     #[allow(clippy::too_many_lines)]
     async fn next_batch(&mut self) -> CollectorResult<Option<CollectedBatch>> {
         if let Some(position) = self.pending_commit.take() {